Aula24-Controle De Acesso

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,11 @@ use Illuminate\Support\Facades\DB;
 class PostsController extends Controller
 {
 
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -78,7 +83,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        return view('posts.edit')->with('post',$post);
+        if(auth()->user()->id !== $post->user_id)
+            return redirect('/posts')->with('error','Acesso não autorizado!');
+        else
+            return view('posts.edit')->with('post',$post);
     }
 
     /**
@@ -114,7 +122,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
-        return redirect('/posts')->with('success','A Postagem foi removida!');
+        $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Acesso não autorizado!');
+        }
+        else{
+            $post->delete();
+            return redirect('/posts')->with('success','A Postagem foi removida!');
+        }    
+        
     }
 }
